# docs/api.yaml
openapi: 3.0.0
info:
  title: DroneAnalytics API
  description: Service for visualizing drone data for discipline "Cyberimmune Systems Software Engineering"
  contact:
    name: "Erkhov A. A. - Admin"
    url: "https://github.com/SashaErkhov"
    email: "sasori654654@gmail.com"
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0

servers:
  - url: https://localhost/api
    description: dev
  - url: https://infopanel.csse.ru/api
    description: prod

tags:
  - name: auth
    description: registration and authentication.
  - name: log
    description: logs.
  - name: analytics
    description: analytics for frontend.

security:
  - BearerAuth: []

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: login and receive access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  $ref: '#/components/schemas/refresh_token'
              required: [refresh_token]
      responses:
        200:
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [auth]
      summary: logout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  $ref: '#/components/schemas/refresh_token'
              required: [refresh_token]
      responses:
        200:
          description: Logs accepted successfully
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/AcceptedResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'

  /log/telemetry:
    post:
      security:
        - ApiKeyAuth: []
      tags: [ log ]
      summary: telemetry of drone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 1000
              items:
                type: object
                additionalProperties: false
                properties:
                  apiVersion:
                    $ref: '#/components/schemas/apiVersion'
                  timestamp:
                    $ref: '#/components/schemas/timestamp'
                  drone:
                    $ref: '#/components/schemas/drone'
                  drone_id:
                    $ref: '#/components/schemas/id'
                  battery:
                    $ref: '#/components/schemas/battery'
                  pitch:
                    $ref: '#/components/schemas/pitch'
                  roll:
                    $ref: '#/components/schemas/roll'
                  course:
                    $ref: '#/components/schemas/course'
                  latitude:
                    $ref: '#/components/schemas/latitude'
                  longitude:
                    $ref: '#/components/schemas/longitude'
                required: [ timestamp, drone, drone_id, latitude, longitude, apiVersion ]
      responses:
        200:
          description: Logs accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        400:
          $ref: '#/components/responses/ValidationError'
        502:
          $ref: '#/components/responses/LogStorageError'
    get:
      security:
        - BearerAuth: []
      tags: [ log ]
      summary: get telemetry logs
      description: Returns telemetry logs sorted by timestamp
      parameters:
        - in: query
          name: limit
          schema:
            $ref: '#/components/schemas/limit'
        - in: query
          name: page
          schema:
            $ref: '#/components/schemas/page'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  properties:
                    timestamp:
                      $ref: '#/components/schemas/timestamp'
                    drone:
                      $ref: '#/components/schemas/drone'
                    drone_id:
                      $ref: '#/components/schemas/id'
                    battery:
                      $ref: '#/components/schemas/battery'
                    pitch:
                      $ref: '#/components/schemas/pitch'
                    roll:
                      $ref: '#/components/schemas/roll'
                    course:
                      $ref: '#/components/schemas/course'
                    latitude:
                      $ref: '#/components/schemas/latitude'
                    longitude:
                      $ref: '#/components/schemas/longitude'
                  required: [ timestamp, drone, drone_id, latitude, longitude ]
        401:
          $ref: '#/components/responses/UnauthorizedError'
        502:
          $ref: '#/components/responses/LogStorageError'
  /log/basic:
    post:
      security:
        - ApiKeyAuth: []
      tags: [ log ]
      summary: the most generalized format of log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 1000
              items:
                $ref: '#/components/schemas/BasicLog'
      responses:
        200:
          description: Logs accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        400:
          $ref: '#/components/responses/ValidationError'
        502:
          $ref: '#/components/responses/LogStorageError'
    get:
      security:
        - BearerAuth: []
      tags: [ log ]
      summary: get basic logs
      description: Returns basic logs sorted by timestamp
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            $ref: '#/components/schemas/limit'
        - in: query
          name: page
          required: false
          schema:
            $ref: '#/components/schemas/page'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BasicLog'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        502:
          $ref: '#/components/responses/LogStorageError'
  /log/event:
    post:
      security:
        - ApiKeyAuth: []
      tags: [ log ]
      summary: event and safety event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              maxItems: 1000
              items:
                type: object
                additionalProperties: false
                properties:
                  apiVersion:
                    $ref: '#/components/schemas/apiVersion'
                  timestamp:
                    $ref: '#/components/schemas/timestamp'
                  event_type:
                    type: string
                    enum: ["event", "safety_event"]
                  service:
                    $ref: '#/components/schemas/service'
                  service_id:
                    $ref: '#/components/schemas/id'
                  severity:
                    $ref: '#/components/schemas/severity'
                  message:
                    $ref: '#/components/schemas/message'
                required: [timestamp, service, service_id, message, apiVersion]
      responses:
        200:
          description: Logs accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        400:
          $ref: '#/components/responses/ValidationError'
        502:
          $ref: '#/components/responses/LogStorageError'
    get:
      security:
        - BearerAuth: []
      tags: [ log ]
      summary: get event logs
      description: Returns event logs sorted by timestamp
      parameters:
        - in: query
          name: limit
          schema:
            $ref: '#/components/schemas/limit'
        - in: query
          name: page
          schema:
            $ref: '#/components/schemas/page'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetEventLog'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        502:
          $ref: '#/components/responses/LogStorageError'
  /log/safety:
    get:
      security:
        - BearerAuth: []
      tags: [ log ]
      summary: get safety events
      description: Returns safety events sorted by timestamp
      parameters:
        - in: query
          name: limit
          schema:
            $ref: '#/components/schemas/limit'
        - in: query
          name: page
          schema:
            $ref: '#/components/schemas/page'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetEventLog'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        502:
          $ref: '#/components/responses/LogStorageError'


components:
  responses:
    ValidationError:
      description: Bad Request - validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    UnauthorizedError:
      description: Unauthorized - authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    LogStorageError:
      description: Bad Gateway - log storage service unavailable or returned an error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'


  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    message:
      type: string
      minLength: 1
      maxLength: 1024
      description: Use only english
    access_token:
      type: string
      minLength: 16
      maxLength: 1024
    refresh_token:
      type: string
      minLength: 16
      maxLength: 1024
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 64
      required: [username, password]
    apiVersion:
      type: string
      description: API version
      default: 1.0.0
      minLength: 5
      maxLength: 8

    severity:
      type: string
      enum: ["debug", "info", "notice", "warning", "error", "critical", "alert", "emergency"]
      description: severity of event - Like syslog

    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: Application-specific error code.
          minimum: 200
          maximum: 599
        message:
          $ref: '#/components/schemas/message'
      required: [code, message]

    timestamp:
      type: integer
      format: int64
      description: Unix timestamp in milliseconds
      minimum: 0

    drone:
      type: string
      description: Type of drone
      enum: ["delivery", "queen", "inspector", "agriculture"]

    service:
      type: string
      description: Type of service
      enum: ["delivery", "queen", "inspector", "agriculture", "GCS", "aggregator", "insurance", "regulator",
             "dronePort", "OrAT_drones", "operator", "SITL", "Gazebo", "infopanel", "registry"]
    battery:
      type: integer
      format: int8
      description: battery level in percents
      minimum: 0
      maximum: 100
    pitch:
      type: number
      format: double
      description: pitch clockwise angle in degrees
      minimum: -90
      maximum: 90
    roll:
      type: number
      format: double
      description: roll clockwise angle in degrees
      minimum: -180
      maximum: 180
    course:
      type: number
      format: double
      description: angle between magnetic north and aircraft axis
      minimum: 0
      maximum: 360
    latitude:
      type: number
      format: double
      description: + north / - south latitude
      minimum: -90
      maximum: 90
    longitude:
      type: number
      format: double
      description: + east / - west longitude
      minimum: -180
      maximum: 180
    id:
      type: integer
      minimum: 1
      maximum: 1000
      format: int64
      description: ID of drone or service. For example, we a have 3 delivery drones, then drone_id will be 1, 2, 3.

    limit:
      type: integer
      minimum: 1
      maximum: 100
      default: 10

    page:
      type: integer
      minimum: 1
      default: 1

    BasicLog:
      type: object
      additionalProperties: false
      properties:
        timestamp:
          $ref: '#/components/schemas/timestamp'
        message:
          $ref: '#/components/schemas/message'
      required: [timestamp, message]
    GetEventLog:
      type: object
      additionalProperties: false
      properties:
        timestamp:
          $ref: '#/components/schemas/timestamp'
        service:
          $ref: '#/components/schemas/service'
        service_id:
          $ref: '#/components/schemas/id'
        severity:
          $ref: '#/components/schemas/severity'
        message:
          $ref: '#/components/schemas/message'
      required: [timestamp, service, service_id, message]

    TokenPair:
      type: object
      properties:
        access_token:
          $ref: '#/components/schemas/access_token'
        refresh_token:
          $ref: '#/components/schemas/refresh_token'
        token_type:
          type: string
          default: "Bearer"
          maxLength: 16
        expires_in:
          type: integer
          format: int64
          description: Access token lifetime in seconds
          minimum: 60
          maximum: 86400
      required: [access_token, refresh_token]
    AcceptedResponse:
      type: object
      properties:
        accepted:
          type: integer
          description: Number of log entries successfully indexed
      required:
        - accepted